moved news to own folder in preperation for child components
